+++ NEWS.md
@@ -1,6 +1,7 @@
 # colorist 0.1.1
 
 - Vignette added
+- Adding an additional lambda parameter to control specificity
 - Bug fixes
 
 # colorist 0.1.0